split memory provider to example types as function
